Break up cluttered class
