+++ util/spec_helper.rb
@@ -39,7 +39,10 @@ require_relative '../pages/page'
 require_relative '../pages/oec/blue_page'
 require_relative '../pages/cal_net_page'
 require_relative '../pages/canvas/canvas_page'
-require_relative '../pages/canvas/canvas_activities_page'
+require_relative '../pages/canvas/canvas_assignments_page'
+require_relative '../pages/canvas/canvas_announce_discuss_page'
+require_relative '../pages/canvas/canvas_grades_page'
+require_relative '../pages/canvas/canvas_groups_page'
 require_relative '../pages/junction/junction_pages'
 require_relative '../pages/junction/api_academics_course_provision_page'
 require_relative '../pages/junction/api_academics_roster_page'
